+++ ui/src/lib/link.js
@@ -41,35 +41,14 @@ function visiblePartners (node) {
 }
 
 function leftPartnersCount (node) {
-  return sideCount(node, -1)
+  const len = visiblePartners(node)
+  return len
 }
 
 function rightPartnersCount (node) {
-  if (node.data.partners && node.data.partners.length === 1) return 0
-  return sideCount(node, 1)
-}
-
-function sideCount (node, direction) {
   const len = visiblePartners(node)
-
-  if (len === 0) return 0
-  if (len === 1) return 1
-
-  const mid = halfwayPoint(len)
-
-  return node.data.partners
-    .filter((d, i) => {
-      var sign = i >= mid ? 1 : -1
-
-      return sign === direction
-    })
-    .length
-}
-
-function halfwayPoint (len) {
-  return len % 2 === 0
-    ? len / 2
-    : Math.round(len / 2) - 1
+  if (node.data.partners && len === 1) return 0
+  return len
 }
 
 // TODO (later): move these settings into vuex where the tree style can be manipulated there